+++ examples/hubert/dataset/__init__.py
@@ -1,12 +1,10 @@
 from .hubert_dataset import (
-    BucketizeSampler,
     CollateFnHubert,
     HuBERTDataSet,
 )
 
 
 __all__ = [
-    "BucketizeSampler",
     "CollateFnHubert",
     "HuBERTDataSet",
 ]
