+++ torchaudio/prototype/datasets/__init__.py
@@ -1,0 +1,5 @@
+from .bucketize_sampler import BucketizeBatchSampler
+
+__all__ = [
+    "BucketizeBatchSampler",
+]
